cleanup yoroi wallets (#2415)
* cleanup yoroi wallets

* better types

* test makeKeys

* Update src/yoroi-wallets/cardano/shelley/ShelleyWallet.ts


---------
