+++ src/SelectedWallet/WalletSelection/WalletSelectionScreen.tsx
@@ -12,10 +12,10 @@ import {useWalletNavigation} from '../../navigation'
 import {COLORS} from '../../theme'
 import {useWalletManager} from '../../WalletManager'
 import {useOpenWallet, useWalletMetas, WALLET_IMPLEMENTATION_REGISTRY, WalletMeta} from '../../yoroi-wallets'
+import * as HASKELL_SHELLEY from '../../yoroi-wallets/cardano/constants/mainnet/constants'
+import * as HASKELL_SHELLEY_TESTNET from '../../yoroi-wallets/cardano/constants/testnet/constants'
 import {InvalidState, NetworkError} from '../../yoroi-wallets/cardano/errors'
 import {isJormungandr} from '../../yoroi-wallets/cardano/networks'
-import * as HASKELL_SHELLEY from '../../yoroi-wallets/cardano/shelley/constants'
-import * as HASKELL_SHELLEY_TESTNET from '../../yoroi-wallets/cardano/shelley-testnet/constants'
 import {useSetSelectedWallet, useSetSelectedWalletMeta} from '../Context'
 import {WalletListItem} from './WalletListItem'
