+++ src/WalletInit/MnemonicShow/MnemonicShowScreen.stories.tsx
@@ -3,7 +3,7 @@ import {storiesOf} from '@storybook/react-native'
 import React from 'react'
 
 import {debugWalletInfo} from '../../features'
-import {NETWORK_ID, WALLET_IMPLEMENTATION_ID} from '../../yoroi-wallets/cardano/shelley-testnet/constants'
+import {NETWORK_ID, WALLET_IMPLEMENTATION_ID} from '../../yoroi-wallets/cardano/constants/testnet/constants'
 import {MnemonicShowScreen} from './MnemonicShowScreen'
 
 storiesOf('MnemonicShowScreen', module).add('Default', () => {
