+++ src/WalletInit/WalletFreshInit/WalletFreshInitScreen.tsx
@@ -9,8 +9,8 @@ import {isNightly} from '../../legacy/config'
 import {WalletInitRouteNavigation} from '../../navigation'
 import {COLORS} from '../../theme'
 import {NetworkId, WALLET_IMPLEMENTATION_REGISTRY, WalletImplementationId} from '../../yoroi-wallets'
-import * as HASKELL_SHELLEY from '../../yoroi-wallets/cardano/shelley/constants'
-import * as HASKELL_SHELLEY_TESTNET from '../../yoroi-wallets/cardano/shelley-testnet/constants'
+import * as HASKELL_SHELLEY from '../../yoroi-wallets/cardano/constants/mainnet/constants'
+import * as HASKELL_SHELLEY_TESTNET from '../../yoroi-wallets/cardano/constants/testnet/constants'
 import {WalletDescription} from '../WalletDescription'
 
 export const WalletFreshInitScreen = () => {
