+++ src/WalletInit/WalletInit/WalletInitScreen.stories.tsx
@@ -3,7 +3,7 @@ import {storiesOf} from '@storybook/react-native'
 import React from 'react'
 
 import {NETWORKS, WALLETS} from '../../yoroi-wallets'
-import {NETWORK_ID, WALLET_IMPLEMENTATION_ID} from '../../yoroi-wallets/cardano/shelley/constants'
+import {NETWORK_ID, WALLET_IMPLEMENTATION_ID} from '../../yoroi-wallets/cardano/constants/mainnet/constants'
 import {WalletInitScreen} from './WalletInitScreen'
 
 storiesOf('WalletInitScreen', module)
