+++ src/WalletInit/WalletInit/WalletInitScreen.tsx
@@ -10,8 +10,8 @@ import globalMessages from '../../i18n/global-messages'
 import {WalletInitRouteNavigation, WalletInitRoutes} from '../../navigation'
 import {COLORS} from '../../theme'
 import {HARDWARE_WALLETS, isByron, isHaskellShelley, NetworkId, WalletImplementationId} from '../../yoroi-wallets'
+import {WALLET_CONFIG_24} from '../../yoroi-wallets/cardano/constants/mainnet/constants'
 import {isJormungandr} from '../../yoroi-wallets/cardano/networks'
-import {WALLET_CONFIG_24} from '../../yoroi-wallets/cardano/shelley/constants'
 import {WalletDescription} from '../WalletDescription'
 import {ExpandableItem} from './ExpandableItem'
