+++ src/yoroi-wallets/cardano/byron/index.ts
@@ -1,0 +1,1 @@
+export * from './ByronWallet'
