--- src/yoroi-wallets/cardano/constants.ts
@@ -1,1 +1,0 @@
-export const HISTORY_REFRESH_TIME = 25000
