+++ src/yoroi-wallets/cardano/index.ts
@@ -1,9 +1,11 @@
 export * from './api'
 export * from './assetUtils'
+export * from './byron'
 export * from './catalyst'
 export * from './chain'
 export * from './formatPath'
 export * from './getMinAda'
+export * from './getWallet'
 export * from './hw'
 export * from './mnemonic'
 export * from './MultiToken'
