+++ src/yoroi-wallets/cardano/processTransactions/index.ts
@@ -1,0 +1,1 @@
+export * from './processTransactions'
