+++ src/yoroi-wallets/cardano/shelley/index.ts
@@ -1,2 +1,3 @@
 export * from './delegationUtils'
 export * from './plate'
+export * from './wallets'
