+++ src/yoroi-wallets/cardano/unsignedTx/unsignedTx.test.ts
@@ -1,8 +1,8 @@
 import BigNumber from 'bignumber.js'
 
-import {YoroiAmounts, YoroiEntries, YoroiMetadata} from '../types'
-import {CardanoTypes} from '.'
-import {toAmounts, toDisplayAddress, toEntries, toMetadata} from './unsignedTx'
+import {YoroiAmounts, YoroiEntries, YoroiMetadata} from '../../types'
+import {CardanoTypes} from '..'
+import {toAmounts, toDisplayAddress, toEntries, toMetadata} from '.'
 
 describe('YoroiUnsignedTx', () => {
   it('toAmounts converts TokenEntry[] to YoroiAmounts', () => {
